test(@vben-core/form-ui): harden benchmark fixtures
